requirements.txt
alembic==1.4.2
asgiref==3.2.7
beautifulsoup4==4.9.1
blinker==1.4
blurhash==1.1.4
boost==0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
decorator==4.4.2
Django==3.0.5
dlib==19.19.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
image==1.5.31
imutils==0.5.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
Mastodon.py==1.5.1
numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==7.1.2
pycparser==2.20
pygame==1.9.6
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-magic==0.4.15
python-socketio==4.5.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.16
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.2.1
xlrd==1.2.0
